+++ emu/aluvop3b.go
@@ -59,6 +59,7 @@ func (u *ALUImpl) runVADDCU32VOP3b(state InstEmuState) {
 	}
 }
 
+//nolint:gocyclo,funlen
 func (u *ALUImpl) runVDIVSCALEF64(state InstEmuState) {
 	sp := state.Scratchpad().AsVOP3B()
 	var i uint
@@ -75,21 +76,21 @@ func (u *ALUImpl) runVDIVSCALEF64(state InstEmuState) {
 		src1 := math.Float64frombits(sp.SRC1[i])
 		src2 := math.Float64frombits(sp.SRC2[i])
 
-		exponentSrc1 := uint64((uint64(sp.SRC1[i]) << 1) >> 53)
-		exponentSrc2 := uint64((uint64(sp.SRC2[i]) << 1) >> 53)
+		exponentSrc1 := (sp.SRC1[i] << 1) >> 53
+		exponentSrc2 := (sp.SRC2[i] << 1) >> 53
 
 		diffExpSrc2Src1 := int64(exponentSrc2) - int64(exponentSrc1)
 
-		fractionSrc1 := uint64((uint64(sp.SRC1[i]) << 12) >> 12)
+		fractionSrc1 := (sp.SRC1[i] << 12) >> 12
 
 		reversedSrc1 := float64(1) / src1
 		src2DivSrc1 := src2 / src1
 
-		exponentRevSrc1 := uint64((uint64(reversedSrc1) << 1) >> 53)
-		fractionRevSrc1 := uint64((uint64(reversedSrc1) << 12) >> 12)
+		exponentRevSrc1 := (uint64(reversedSrc1) << 1) >> 53
+		fractionRevSrc1 := (uint64(reversedSrc1) << 12) >> 12
 
-		exponentSrc2DivSrc1 := uint64((uint64(src2DivSrc1) << 1) >> 53)
-		fractionSrc2DivSrc1 := uint64((uint64(src2DivSrc1) << 12) >> 12)
+		exponentSrc2DivSrc1 := (uint64(src2DivSrc1) << 1) >> 53
+		fractionSrc2DivSrc1 := (uint64(src2DivSrc1) << 12) >> 12
 
 		if src2 == 0 || src1 == 0 {
 			sp.DST[i] = 0x7FFFFFFFFFFFFFFF // NaN
